banner uploader working, making successful PUT req to our backend
