+++ app/src/styles/scss/vendor_profile.module.scss
@@ -54,7 +54,7 @@
         z-index: 2;
         margin-top: 5px;
         padding: 10px;
-        height: 0px;
+        height: fit-content;
          p{        
           cursor: pointer;
           position: relative;
